+++ packages/twenty-server/src/modules/messaging/services/gmail-messages-import/gmail-messages-import-v2.service.ts
@@ -1,11 +1,6 @@
-import { Inject, Injectable, Logger } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 
-import { DataSource, EntityManager } from 'typeorm';
-
-import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
-import { ConnectedAccountRepository } from 'src/modules/connected-account/repositories/connected-account.repository';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
-import { MessageChannelRepository } from 'src/modules/messaging/repositories/message-channel.repository';
 import { FetchMessagesByBatchesService } from 'src/modules/messaging/services/fetch-messages-by-batches/fetch-messages-by-batches.service';
 import {
   MessageChannelWorkspaceEntity,
@@ -16,20 +11,11 @@ import { InjectCacheStorage } from 'src/engine/integrations/cache-storage/decora
 import { CacheStorageNamespace } from 'src/engine/integrations/cache-storage/types/cache-storage-namespace.enum';
 import { CacheStorageService } from 'src/engine/integrations/cache-storage/cache-storage.service';
 import { GMAIL_USERS_MESSAGES_GET_BATCH_SIZE } from 'src/modules/messaging/constants/gmail-users-messages-get-batch-size.constant';
-import { WorkspaceDataSourceService } from 'src/engine/workspace-datasource/workspace-datasource.service';
-import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
-import { MessageQueueService } from 'src/engine/integrations/message-queue/services/message-queue.service';
 import { GMAIL_ONGOING_SYNC_TIMEOUT } from 'src/modules/messaging/constants/gmail-ongoing-sync-timeout.constant';
-import { MessageParticipantService } from 'src/modules/messaging/services/message-participant/message-participant.service';
-import { MessageService } from 'src/modules/messaging/services/message/message.service';
-import { ParticipantWithMessageId } from 'src/modules/messaging/types/gmail-message';
-import {
-  CreateCompanyAndContactJobData,
-  CreateCompanyAndContactJob,
-} from 'src/modules/connected-account/auto-companies-and-contacts-creation/jobs/create-company-and-contact.job';
 import { GmailMessagesImportService } from 'src/modules/messaging/services/gmail-messages-import/gmail-messages-import.service';
 import { SetMessageChannelSyncStatusService } from 'src/modules/messaging/services/set-message-channel-sync-status/set-message-channel-sync-status.service';
 import { ObjectRecord } from 'src/engine/workspace-manager/workspace-sync-metadata/types/object-record';
+import { SaveMessagesAndEnqueueContactCreationService } from 'src/modules/messaging/services/gmail-messages-import/save-messages-and-enqueue-contact-creation.service';
 
 @Injectable()
 export class GmailMessagesImportV2Service {
@@ -37,18 +23,10 @@ export class GmailMessagesImportV2Service {
 
   constructor(
     private readonly fetchMessagesByBatchesService: FetchMessagesByBatchesService,
-    @InjectObjectMetadataRepository(ConnectedAccountWorkspaceEntity)
-    private readonly connectedAccountRepository: ConnectedAccountRepository,
-    @InjectObjectMetadataRepository(MessageChannelWorkspaceEntity)
-    private readonly messageChannelRepository: MessageChannelRepository,
     @InjectCacheStorage(CacheStorageNamespace.Messaging)
     private readonly cacheStorage: CacheStorageService,
-    private readonly workspaceDataSourceService: WorkspaceDataSourceService,
-    @Inject(MessageQueue.messagingQueue)
-    private readonly messageQueueService: MessageQueueService,
-    private readonly messageService: MessageService,
-    private readonly messageParticipantService: MessageParticipantService,
     private readonly setMessageChannelSyncStatusService: SetMessageChannelSyncStatusService,
+    private readonly saveMessagesAndEnqueueContactCreationService: SaveMessagesAndEnqueueContactCreationService,
   ) {}
 
   async processMessageBatchImport(
@@ -99,11 +77,6 @@ export class GmailMessagesImportV2Service {
       return;
     }
 
-    const workspaceDataSource =
-      await this.workspaceDataSourceService.connectToWorkspaceDataSource(
-        workspaceId,
-      );
-
     const messageQueries = createQueriesFromMessageIds(messageIdsToFetch);
 
     try {
@@ -124,14 +97,12 @@ export class GmailMessagesImportV2Service {
         return [];
       }
 
-      const participantsWithMessageId =
-        await this.saveMessagesAndParticipantsAndReturnParticipantsWithMessageId(
-          messagesToSave,
-          connectedAccount,
-          messageChannel,
-          workspaceId,
-          workspaceDataSource,
-        );
+      await this.saveMessagesAndEnqueueContactCreationService.saveMessagesAndEnqueueContactCreationJob(
+        messagesToSave,
+        messageChannel,
+        connectedAccount,
+        workspaceId,
+      );
 
       if (messageIdsToFetch.length < GMAIL_USERS_MESSAGES_GET_BATCH_SIZE) {
         await this.setMessageChannelSyncStatusService.setCompletedStatus(
@@ -150,21 +121,6 @@ export class GmailMessagesImportV2Service {
 
         this.logger.log(
           `Messaging import for workspace ${workspaceId} and account ${connectedAccount.id} done with more messages to import.`,
-        );
-      }
-
-      if (messageChannel.isContactAutoCreationEnabled) {
-        const contactsToCreate = participantsWithMessageId.filter(
-          (participant) => participant.shouldCreateContact,
-        );
-
-        await this.messageQueueService.add<CreateCompanyAndContactJobData>(
-          CreateCompanyAndContactJob.name,
-          {
-            workspaceId,
-            connectedAccountHandle: connectedAccount.handle,
-            contactsToCreate,
-          },
         );
       }
     } catch (error) {
@@ -186,53 +142,5 @@ export class GmailMessagesImportV2Service {
         `Error fetching messages for ${connectedAccount.id} in workspace ${workspaceId}: ${error.message}`,
       );
     }
-  }
-
-  public async saveMessagesAndParticipantsAndReturnParticipantsWithMessageId(
-    messagesToSave: any[],
-    connectedAccount: ObjectRecord<ConnectedAccountWorkspaceEntity>,
-    messageChannel: ObjectRecord<MessageChannelWorkspaceEntity>,
-    workspaceId: string,
-    workspaceDataSource: DataSource,
-  ) {
-    const participantsWithMessageId = await workspaceDataSource?.transaction(
-      async (transactionManager: EntityManager) => {
-        const messageExternalIdsAndIdsMap =
-          await this.messageService.saveMessagesWithinTransaction(
-            messagesToSave,
-            connectedAccount,
-            messageChannel.id,
-            workspaceId,
-            transactionManager,
-          );
-
-        const participantsWithMessageId: (ParticipantWithMessageId & {
-          shouldCreateContact: boolean;
-        })[] = messagesToSave.flatMap((message) => {
-          const messageId = messageExternalIdsAndIdsMap.get(message.externalId);
-
-          return messageId
-            ? message.participants.map((participant) => ({
-                ...participant,
-                messageId,
-                shouldCreateContact:
-                  messageChannel.isContactAutoCreationEnabled &&
-                  message.participants.find((p) => p.role === 'from')
-                    ?.handle === connectedAccount.handle,
-              }))
-            : [];
-        });
-
-        await this.messageParticipantService.saveMessageParticipants(
-          participantsWithMessageId,
-          workspaceId,
-          transactionManager,
-        );
-
-        return participantsWithMessageId;
-      },
-    );
-
-    return participantsWithMessageId;
   }
 }
