+++ packages/twenty-server/src/modules/messaging/services/gmail-messages-import/gmail-messages-import.module.ts
@@ -6,6 +6,7 @@ import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/s
 import { FetchMessagesByBatchesModule } from 'src/modules/messaging/services/fetch-messages-by-batches/fetch-messages-by-batches.module';
 import { GmailMessagesImportV2Service } from 'src/modules/messaging/services/gmail-messages-import/gmail-messages-import-v2.service';
 import { GmailMessagesImportService } from 'src/modules/messaging/services/gmail-messages-import/gmail-messages-import.service';
+import { SaveMessagesAndEnqueueContactCreationService } from 'src/modules/messaging/services/gmail-messages-import/save-messages-and-enqueue-contact-creation.service';
 import { MessageParticipantModule } from 'src/modules/messaging/services/message-participant/message-participant.module';
 import { MessageModule } from 'src/modules/messaging/services/message/message.module';
 import { SetMessageChannelSyncStatusModule } from 'src/modules/messaging/services/set-message-channel-sync-status/set-message-channel-sync-status.module';
@@ -24,7 +25,11 @@ import { MessageChannelWorkspaceEntity } from 'src/modules/messaging/standard-ob
     MessageParticipantModule,
     SetMessageChannelSyncStatusModule,
   ],
-  providers: [GmailMessagesImportService, GmailMessagesImportV2Service],
+  providers: [
+    GmailMessagesImportService,
+    GmailMessagesImportV2Service,
+    SaveMessagesAndEnqueueContactCreationService,
+  ],
   exports: [GmailMessagesImportService, GmailMessagesImportV2Service],
 })
 export class GmailMessagesImportModule {}
